issues fixed
for the mobile translation, the icon fits beside the light-dark mode. For the vote proposal modal, it does appear please check again as the changes that i did previously was to place the text inside it inside the translation instead of being hard-coded. flags are deleted as there is no need for them (maybe later we can add this feature). Steem stats also fixed as it was replaced by mistake and the image in the home page is replaced with a working one.